# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WorkerProfileStatus = typeof WorkerProfileStatus[keyof typeof WorkerProfileStatus];


export const WorkerProfileStatus = {
  pending: 'pending',
  active: 'active',
  paused: 'paused',
} as const;

export type WorkerProfileAvailability = typeof WorkerProfileAvailability[keyof typeof WorkerProfileAvailability];


export const WorkerProfileAvailability = {
  available: 'available',
  away: 'away',
} as const;

export interface WorkerProfile {
  id: string;
  name: string;
  email: string;
  city?: string;
  country?: string;
  phoneNumber?: string;
  education?: string;
  yearOfPass?: number;
  gpa?: string;
  avatarInitials?: string;
  status: WorkerProfileStatus;
  availability: WorkerProfileAvailability;
  timezone?: string;
  skills: string[];
  languages: string[];
  experience?: string;
  weeklyHours?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  onboardingProgress: number;
  internshipTermsAccepted: boolean;
  /** @nullable */
  internshipTermsAcceptedAt?: string | null;
}

export type WorkerProfileUpdateAvailability = typeof WorkerProfileUpdateAvailability[keyof typeof WorkerProfileUpdateAvailability];


export const WorkerProfileUpdateAvailability = {
  available: 'available',
  away: 'away',
} as const;

export interface WorkerProfileUpdate {
  name?: string;
  city?: string;
  country?: string;
  phoneNumber?: string;
  education?: string;
  yearOfPass?: number;
  gpa?: string;
  availability?: WorkerProfileUpdateAvailability;
  timezone?: string;
  skills?: string[];
  languages?: string[];
  experience?: string;
  weeklyHours?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  onboardingProgress?: number;
  internshipTermsAccepted?: boolean;
}

export type TaskDifficulty = typeof TaskDifficulty[keyof typeof TaskDifficulty];


export const TaskDifficulty = {
  starter: 'starter',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  available: 'available',
  claimed: 'claimed',
  submitted: 'submitted',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type TaskTaskType = typeof TaskTaskType[keyof typeof TaskTaskType];


export const TaskTaskType = {
  standard: 'standard',
  video_screening: 'video_screening',
} as const;

export interface Task {
  id: string;
  title: string;
  description: string;
  category: string;
  difficulty: TaskDifficulty;
  estimatedMinutes: number;
  payout: number;
  status: TaskStatus;
  taskType: TaskTaskType;
  requiredSkills: string[];
  submissionsCount: number;
  createdAt: string;
  /** @nullable */
  dueAt?: string | null;
  /** @nullable */
  claimedAt?: string | null;
}

export type TaskInputDifficulty = typeof TaskInputDifficulty[keyof typeof TaskInputDifficulty];


export const TaskInputDifficulty = {
  starter: 'starter',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  category: string;
  difficulty: TaskInputDifficulty;
  /** @minimum 1 */
  estimatedMinutes: number;
  /** @minimum 0 */
  payout: number;
  requiredSkills: string[];
}

export interface TaskSubmissionInput {
  /** @minLength 1 */
  response: string;
  notes?: string;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  in_progress: 'in_progress',
  submitted: 'submitted',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Submission {
  id: string;
  taskId: string;
  taskTitle: string;
  workerName?: string;
  response?: string;
  /** @nullable */
  notes?: string | null;
  status: SubmissionStatus;
  submittedAt: string;
  payout: number;
  /** @nullable */
  reviewerNote?: string | null;
  /** @nullable */
  videoUrl?: string | null;
}

export interface VideoScreeningInput {
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  contentType: string;
  /**
     * @minimum 1
     * @maximum 180
     */
  durationSeconds: number;
}

export type WorkerScreeningStatus = typeof WorkerScreeningStatus[keyof typeof WorkerScreeningStatus];


export const WorkerScreeningStatus = {
  pending: 'pending',
  submitted: 'submitted',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface WorkerScreening {
  task: Task;
  status: WorkerScreeningStatus;
  /** @nullable */
  submissionId: string | null;
  /** @nullable */
  videoUrl: string | null;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ReviewItem = Submission & {
  workerEmail: string;
  category: string;
};

export type ReviewInputStatus = typeof ReviewInputStatus[keyof typeof ReviewInputStatus];


export const ReviewInputStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ReviewInput {
  status: ReviewInputStatus;
  reviewerNote?: string;
}

export type WorkerDashboardStats = {
  availableTasks: number;
  completedTasks: number;
  pendingReview: number;
  earned: number;
};

export interface WorkerDashboard {
  profile: WorkerProfile;
  stats: WorkerDashboardStats;
  featuredTasks: Task[];
  recentSubmissions: Submission[];
}

export type DispatcherWorkerAvailability = typeof DispatcherWorkerAvailability[keyof typeof DispatcherWorkerAvailability];


export const DispatcherWorkerAvailability = {
  available: 'available',
  away: 'away',
} as const;

export type DispatcherWorkerStatus = typeof DispatcherWorkerStatus[keyof typeof DispatcherWorkerStatus];


export const DispatcherWorkerStatus = {
  pending: 'pending',
  active: 'active',
  paused: 'paused',
} as const;

export interface DispatcherWorker {
  id: string;
  name: string;
  email: string;
  availability: DispatcherWorkerAvailability;
  status: DispatcherWorkerStatus;
  skills: string[];
  /** @nullable */
  activeTaskTitle?: string | null;
  /** @nullable */
  activeTaskStatus?: string | null;
  /** @nullable */
  lastSubmissionAt?: string | null;
  completedTasks: number;
  pendingReview: number;
}

export type DispatcherTask = Task & ({
  /** @nullable */
  claimedByName: string | null;
  /** @nullable */
  claimedByEmail: string | null;
});

export type DispatcherOverviewStats = {
  totalWorkers: number;
  availableWorkers: number;
  activeTasks: number;
  tasksInReview: number;
  completedToday: number;
  completionRate: number;
};

export interface DispatcherOverview {
  stats: DispatcherOverviewStats;
  workers: DispatcherWorker[];
  tasks: DispatcherTask[];
}

export type ListTasksParams = {
status?: ListTasksStatus;
category?: string;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  available: 'available',
  claimed: 'claimed',
  submitted: 'submitted',
  all: 'all',
} as const;