# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions/{submissionId}/video:
    get:
      operationId: getSubmissionVideo
      tags: [worker]
      summary: Stream a private video submission
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Private video stream
          content:
            video/webm:
              schema:
                type: string
                format: binary
        "403":
          description: Not allowed to view this video
        "404":
          description: Video submission not found
  /admin/tasks:
    post:
      operationId: createTask
      tags: [admin]
      summary: Publish a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Published task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /admin/submissions/{submissionId}/review:
    post:
      operationId: reviewSubmission
      tags: [admin]
      summary: Review a worker submission
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Reviewed submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [Storage]
      summary: Request a presigned URL for a private upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid upload metadata
        "401":
          description: Authentication required
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [Storage]
      summary: Stream a private object owned by the current user
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Private object stream
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "403":
          description: Not allowed to view this object
        "404":
          description: Object not found
components:
...
  schemas:
...
    Task:
      type: object
      required:
        - id
        - title
        - description
        - category
        - difficulty
        - estimatedMinutes
        - payout
        - status
        - taskType
        - requiredSkills
        - submissionsCount
        - createdAt
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [starter, intermediate, advanced]
        estimatedMinutes:
          type: number
        payout:
          type: number
        status:
          type: string
          enum: [available, claimed, submitted, approved, rejected]
        taskType:
          type: string
          enum: [standard, video_screening]
        requiredSkills:
          type: array
          items:
            type: string
        submissionsCount:
          type: number
        createdAt:
          type: string
          format: date-time
        dueAt:
          type: ["string", "null"]
          format: date-time
        claimedAt:
          type: ["string", "null"]
          format: date-time
    TaskInput:
      type: object
      required: [title, description, category, difficulty, estimatedMinutes, payout, requiredSkills]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        category:
          type: string
        difficulty:
          type: string
          enum: [starter, intermediate, advanced]
        estimatedMinutes:
          type: number
          minimum: 1
        payout:
          type: number
          minimum: 0
        requiredSkills:
          type: array
          items:
            type: string
...
    Submission:
      type: object
      required:
        - id
        - taskId
        - taskTitle
        - status
        - submittedAt
        - payout
      properties:
        id:
          type: string
        taskId:
          type: string
        taskTitle:
          type: string
        workerName:
          type: string
        response:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, submitted, approved, rejected]
        submittedAt:
          type: string
          format: date-time
        payout:
          type: number
        reviewerNote:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...
    ReviewInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewerNote:
          type: string
...